modularized local strategies and secrets
