+++ auth/passport.js
@@ -1,56 +1,14 @@
 /**
  * Created by abhishekyadav on 26/06/17.
  */
-var passport = require('passport')
-    , OAuth2Strategy = require('passport-oauth').OAuth2Strategy;
-const LocalStrategy = require('passport-local').Strategy;
+var passport = require('passport');
+var oAuthStrategy = require('./strategies/oAuth/oAuthStrategy');
+const localStrategy = require('./strategies/local/localStrategy');
 const models = require('../db/models');
 const db = require('../db');
-var http = require('http');
-var randtoken = require('rand-token');
-const axios = require('axios')
 
 
-var options = {
-    uri: 'https://account.codingblocks.com/api/users/me',
-    method: 'GET'
-};
-
-
-passport.use(new LocalStrategy(
-    function (username, password, done) {
-        models.UserLocal.findOne({
-            where: {
-                username: username
-            },
-            include: [{
-                model: models.Users,
-                include: [{
-                    model: models.Students
-                },
-                    {
-                        model: models.Teachers
-                    }]
-            }]
-        }).then(function (data) {
-            if (!data) {
-                return done(null, false, {message: "Incorrect Username"});
-            }
-            else {
-                console.log("hi")
-                db.actions.validatePassword.validateLocalPassword(data.dataValues, password, done, function (user, isValid, callback) {
-                    if (isValid) {
-                        callback(null, user);
-                    }
-                    else {
-                        callback(null, false, {message: 'Invalid Password'});
-                    }
-                });
-            }
-        }).catch(function (err) {
-            if (err) return done(err);
-        });
-    }));
+passport.use(localStrategy);
 
 passport.serializeUser(function (user, done) {
     if (!user.val) {
@@ -84,71 +42,7 @@ passport.deserializeUser(function (obj, done) {
 });
 
 
-passport.use('oauth-cb', new OAuth2Strategy({
-        authorizationURL: 'https://account.codingblocks.com/oauth/authorize',
-        tokenURL: 'https://account.codingblocks.com/oauth/token',
-        clientID: '4096593676',
-        clientSecret: '1WHWF8RW1DCk7aF1kAAVPAF1LKdDsGrTl9SyKOt4S8bfBpK9XnBttA0SegE8ptfY',
-        callbackURL: 'http://localhost:4000/users/login/cb/callback'
-    },
-    function (accessToken, refreshToken, profile, done) {
-        models.AuthToken.findOrCreate(
-            {
-                where: {
-                    accesstoken: accessToken
-                },
-                defaults: {
-                    accesstoken: accessToken,
-                    clientoken: randtoken.generate(16),
-                    user: {}
-                },
-                include: [models.Users]
-            }
-        ).then(function (user) {
-            axios.get('https://account.codingblocks.com/api/users/me', {
-                headers: {
-                    'Authorization': 'Bearer ' + user[0].dataValues.accesstoken
-                }
-            }).then(function (response) {
-                user[0].dataValues.name = response.data.firstname + ' ' + response.data.lastname
-                user[0].dataValues.email = response.data.email
-                models.Students.findOne({
-                    where : {
-                        userId : user[0].dataValues.user.id
-                    }
-                }).then(function (data) {
-                    if(!data){
-                        models.Teachers.findOne({
-                            where : {
-                                userId : user[0].dataValues.user.id
-                            }
-                        }).then(function (data) {
-                            if(!data){
-                                console.log('jcscljshcihaslca')
-                                user[0].dataValues.val = true;
-                            }
-                            else{
-                                user[0].dataValues.val = false;
-                                user[0].dataValues.user.dataValues.student = false;
-                            }
-                            done(null, user[0].dataValues);
-                        })
-                    }
-                    else{
-                        user[0].dataValues.val = false;
-                        user[0].dataValues.user.dataValues.student=true
-                        done(null, user[0].dataValues);
-                    }
-
-                });
-            }).catch(function (err) {
-                if (err) throw err
-            })
-        }).catch(function (err) {
-            done(null, false);
-        })
-    }
-));
+passport.use('oauth-cb',oAuthStrategy );
