+++ secrets.json
@@ -1,0 +1,6 @@
+{
+
+  "clientID": "4096593676",
+  "clientSecret": "1WHWF8RW1DCk7aF1kAAVPAF1LKdDsGrTl9SyKOt4S8bfBpK9XnBttA0SegE8ptfY"
+
+}
